# btc/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is BUTTON Node API responseModels documentation",
        "title": "Swagger BUTTON Node API",
        "contact": {
            "name": "API Support",
            "email": "nk"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/btc/balance/{address}": {
            "get": {
                "description": "return balance of account in BTC for specific node",
                "produces": [
                    "application/json"
                ],
                "summary": "BTC balance of account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.BalanceResponse"
                            }
                        }
                    }
                }
            }
        },
        "/btc/balances": {
            "post": {
                "description": "return balances of accounts in BTC",
                "produces": [
                    "application/json"
                ],
                "summary": "BTC balance of accounts by list",
                "parameters": [
                    {
                        "description": "addressesArray",
                        "name": "addressesArray",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.BalancesResponse"
                            }
                        }
                    }
                }
            }
        },
        "/btc/bestTransactionFee": {
            "get": {
                "description": "return Amount of BTC that you need to send a transaction",
                "produces": [
                    "application/json"
                ],
                "summary": "return Amount of BTC that you need to send a transaction",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.TransactionFeeResponse"
                            }
                        }
                    }
                }
            }
        },
        "/btc/transactionFee": {
            "get": {
                "description": "return BTC fee",
                "produces": [
                    "application/json"
                ],
                "summary": "BTC fee",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.TransactionFeeResponse"
                            }
                        }
                    }
                }
            }
        },
        "/btc/utxo/{address}": {
            "get": {
                "description": "return UTXO of account",
                "produces": [
                    "application/json"
                ],
                "summary": "BTC UTXO of account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.UTXOResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "responses.BalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number",
                    "example": 0
                }
            }
        },
        "responses.BalancesResponse": {
            "type": "object",
            "properties": {
                "balances": {
                    "type": "object"
                }
            }
        },
        "responses.TransactionFeeResponse": {
            "type": "object",
            "properties": {
                "fee": {
                    "type": "number",
                    "example": 0
                }
            }
        },
        "responses.UTXO": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "confirmations": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "satoshis": {
                    "type": "integer"
                },
                "scriptPubKey": {
                    "type": "string"
                },
                "txid": {
                    "type": "string"
                },
                "vout": {
                    "type": "integer"
                }
            }
        },
        "responses.UTXOResponse": {
            "type": "object",
            "properties": {
                "utxo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.UTXO"
                    }
                }
            }
        }
    }
}